// src/lib/api/generated/models/linkCategory.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Thisiscz API
 * Thisiscz 后端 API 文档
 * OpenAPI spec version: v1
 */

export type LinkCategory = (typeof LinkCategory)[keyof typeof LinkCategory]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LinkCategory = {
  life: 'life',
  work: 'work',
  crypto: 'crypto',
  sports: 'sports',
  movies: 'movies',
} as const